feat(sdk): add messageHistory to hitl interface
